模型相关启用、禁用、构建等增加验证

+++ ioc-olap-svc/src/main/java/com/openjava/olap/api/OlapModelingAction.java
@@ -10,6 +10,7 @@ import com.openjava.olap.service.*;
 import com.openjava.olap.vo.CubeListVo;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+import net.sf.ehcache.transaction.xa.EhcacheXAException;
 import org.apache.commons.lang3.StringUtils;
 import org.ljdp.component.exception.APIException;
 import org.ljdp.component.sequence.ConcurrentSequence;
@@ -662,6 +663,34 @@ public class OlapModelingAction extends BaseAction {
     @RequestMapping(value = "/build", method = RequestMethod.PUT)
     @Security(session = true)
     public void build(String cubeName, Long start, Long end, @RequestBody OlapTimingrefresh timingrefresh) throws APIException {
+        //增量全量验证
+        CubeDescDataMapper cubeDescDataMapper = cubeHttpClient.desc(cubeName);
+        if (cubeDescDataMapper == null) {
+            throw new APIException(400, cubeName + "模型已被删除！");
+        }
+        ModelsDescDataMapper modelsDescDataMapper = modelHttpClient.entity(cubeDescDataMapper.model_name);
+        if (modelsDescDataMapper == null) {
+            throw new APIException(400, "模型已被损坏！");
+        }//增量
+        else if (modelsDescDataMapper.partition_desc != null && StringUtils.isNotBlank(modelsDescDataMapper.partition_desc.partition_date_column)) {
+            if (start == 0 || end == 0) {
+                throw new APIException(400, "开始及结束时间不能为0！");
+            }
+            List<CubeHbaseMapper> hbases = cubeHttpClient.hbase(cubeName);
+            if (hbases.stream().filter(p -> (start >= p.getDateRangeStart() && start <= p.getDateRangeEnd()) || (end >= p.getDateRangeStart() && end <= p.getDateRangeEnd()))
+                    .count() > 0) {
+                throw new APIException(400, "构建的时间区间不能重叠！");
+            }
+        }//全量
+        else {
+            if (start != 0 || end != 0) {
+                throw new APIException("全量构建参数错误！");
+            }
+            List<CubeHbaseMapper> hbases = cubeHttpClient.hbase(cubeName);
+            if (hbases.stream().filter(p -> p.getSegmentStatus().equals("NEW")).count() > 0) {
+                throw new APIException(400, "已存在正在构建或失败的作业！");
+            }
+        }
         OaUserVO userVO = (OaUserVO) SsoContext.getUser();
         timingrefresh.setUpdateId(Long.parseLong(userVO.getUserId()));
         timingrefresh.setUpdateName(userVO.getUserAccount());
@@ -675,7 +704,23 @@ public class OlapModelingAction extends BaseAction {
     @RequestMapping(value = "/refresh", method = RequestMethod.PUT)
     @Security(session = true)
     public void refresh(String cubeName, Long startTime, Long endTime) throws Exception {
+        CubeMapper cubeMapper;
+        try {
+            cubeMapper = cubeHttpClient.get(cubeName);
+        } catch (Exception ex) {
+            throw new APIException(400, cubeName + "模型已被删除！");
+        }
+
+        if (startTime == 0 || endTime == 0) {
+            throw new APIException(400, "开始或结束时间不能等于0！");
+        }
+
+        if (cubeMapper.getSegments().stream().filter(p -> p.date_range_start.equals(startTime) && p.date_range_end.equals(endTime)).count() == 0) {
+            throw new APIException(400, "没有找到该时间区间的块！");
+        }
+
         cubeHttpClient.refresh(cubeName, startTime, endTime);
+
     }
 
 
@@ -683,7 +728,18 @@ public class OlapModelingAction extends BaseAction {
     @RequestMapping(value = "/merge", method = RequestMethod.PUT)
     @Security(session = true)
     public void merge(String cubeName, Long start, Long end) throws Exception {
-        cubeHttpClient.build(cubeName, start, end);
+        CubeMapper cubeMapper;
+        try {
+            cubeMapper = cubeHttpClient.get(cubeName);
+        } catch (Exception ex) {
+            throw new APIException(400, cubeName + "模型已被删除！");
+        }
+
+        if (cubeMapper.getSegments().stream().filter(p -> p.date_range_start >= start && p.date_range_end <= end).count() < 2) {
+            throw new APIException(400, "合并的块最少需要2个！");
+        }
+
+        cubeHttpClient.merge(cubeName, start, end);
     }
 
 
@@ -691,9 +747,20 @@ public class OlapModelingAction extends BaseAction {
     @RequestMapping(value = "/disable", method = RequestMethod.PUT)
     @Security(session = true)
     public void disable(String cubeName) throws APIException {
-        cubeHttpClient.disable(cubeName);
+        CubeMapper cubeMapper;
+        try {
+            cubeMapper = cubeHttpClient.get(cubeName);
+        } catch (Exception ex) {
+            throw new APIException(400, cubeName + "模型已被删除！");
+        }
+        if (cubeMapper.getStatus().equals("READY")) {
+            cubeHttpClient.disable(cubeName);
+        }
         //修改olap分析的cube表状态值为不可用
         OlapCube olapCube = olapCubeService.findTableInfo(cubeName);
+        if (olapCube == null) {
+            throw new APIException(400, "模型数据缺失！");
+        }
         olapCube.setIsNew(false);
         olapCube.setFlags(0);
         olapCubeService.doSave(olapCube);
@@ -704,12 +771,29 @@ public class OlapModelingAction extends BaseAction {
     @RequestMapping(value = "/enable", method = RequestMethod.PUT)
     @Security(session = true)
     public void enable(String cubeName) throws APIException {
-        cubeHttpClient.enable(cubeName);
-        //修改olap分析的cube表状态值为不可用
-        OlapCube olapCube = olapCubeService.findTableInfo(cubeName);
-        olapCube.setIsNew(false);
-        olapCube.setFlags(1);
-        olapCubeService.doSave(olapCube);
+        CubeMapper cubeMapper;
+        try {
+            cubeMapper = cubeHttpClient.get(cubeName);
+        } catch (Exception ex) {
+            throw new APIException(400, cubeName + "模型已被删除！");
+        }
+        if (cubeMapper.getSegments().stream().filter(p -> p.getStatus().equals("READY")).count() == 0) {
+            throw new APIException(400, "没有构建成功的块！");
+        }
+        //如果不是这两个状态证明模型损坏
+        if (!cubeMapper.getStatus().equals("DISABLED") && !cubeMapper.getStatus().equals("READY")) {
+            throw new APIException(400, "模型损坏！");
+        }
+        if (cubeMapper.getStatus().equals("DISABLED")) {
+            cubeHttpClient.enable(cubeName);
+            OlapCube olapCube = olapCubeService.findTableInfo(cubeName);
+            if (olapCube == null) {
+                throw new APIException(400, "模型数据缺失！");
+            }
+            olapCube.setIsNew(false);
+            olapCube.setFlags(1);
+            olapCubeService.doSave(olapCube);
+        }
     }
 
     @ApiOperation(value = "立方体:复制")
@@ -717,16 +801,14 @@ public class OlapModelingAction extends BaseAction {
     @Security(session = true)
     public void clone(String cubeName, String cubeNameClone) throws APIException {
         OaUserVO userVO = (OaUserVO) SsoContext.getUser();
-        Date date = new Date();
-
         CubeDescDataMapper cube = cubeHttpClient.desc(cubeNameClone);
         if (cube != null) {
-            throw new APIException(400, "已存在该立方体名称！");
+            throw new APIException(400, "已存在该模型名称！");
         }
         //拿到OLAP_CUBE表数据
         OlapCube olapCubeEntity = olapCubeService.findTableInfo(cubeName);
         if (olapCubeEntity == null) {
-            throw new APIException(400, "立方体数据缺失！");
+            throw new APIException(400, "模型数据缺失！");
         }
         //进行克隆
         cubeHttpClient.clone(cubeName, cubeNameClone, userVO.getUserId());
@@ -751,16 +833,18 @@ public class OlapModelingAction extends BaseAction {
     @RequestMapping(value = "/deleteCube", method = RequestMethod.DELETE)
     @Security(session = true)
     public void deleteCube(String cubeName) throws APIException {
-        OaUserVO userVO = (OaUserVO) SsoContext.getUser();
-        //判断该立方体是否还存在job数据,如果存在则不能删除.
-        List<JobsMapper> jobList = Arrays.asList(jobsHttpClient.list(100000L, 0L, userVO.getUserId(), cubeName));
-        for (JobsMapper jobs : jobList) {
-            if (jobs.getName().equals(cubeName)) {
-                throw new APIException(400, "请先删除构建数据！");
-            }
+        CubeMapper cubeMapper = null;
+        try {
+            cubeMapper = cubeHttpClient.get(cubeName);
+        } catch (Exception ex) {
         }
-        //删除立方体
-        cubeHttpClient.delete(cubeName);
+        if (cubeMapper != null) {
+            if (cubeMapper.getSegments().stream().filter(p -> p.getStatus().equals("NEW")).count() > 0) {
+                throw new APIException(400, "存在正在构建或失败的作业！");
+            }
+            //删除立方体
+            cubeHttpClient.delete(cubeName);
+        }
         //删除与该立方体相关数据
         OlapCube olapCube = olapCubeService.findTableInfo(cubeName);
         if (olapCube != null) {
